increase track contrast, show importbutton + spinner on late import
Fixes #724
FIxes #723
Fixes #281
