+++ lib/app.dart
@@ -1,3 +1,3 @@
-export 'src/app/app.dart';
 export 'src/app/app_model.dart';
+export 'src/app/view/app.dart';
 export 'src/app/view/splash_screen.dart';
